refactor: merge into ffmpeg module
