--- cyberdrop_dl/utils/ffprobe.py
@@ -1,239 +1,0 @@
-from __future__ import annotations
-
-import asyncio
-import json
-import subprocess
-from dataclasses import asdict, dataclass, field
-from datetime import timedelta
-from fractions import Fraction
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, NamedTuple, Required, Self, TypedDict, cast
-
-from multidict import CIMultiDict
-from yarl import URL
-
-from cyberdrop_dl.utils.logger import log_debug
-from cyberdrop_dl.utils.utilities import get_valid_dict
-
-
-class Duration(NamedTuple):
-    # "00:03:48.250000000"
-    days: int = 0
-    hours: int = 0
-    minutes: int = 0
-    seconds: int = 0
-    microseconds: int = 0
-
-    @staticmethod
-    def parse(duration: float | str) -> float:
-        if isinstance(duration, float | int):
-            return duration
-        try:
-            return float(duration)
-        except (ValueError, TypeError):
-            pass
-        days, _, other_parts = duration.partition(" ")
-        if days:
-            days = "".join(c for c in days if c.isdigit())
-        else:
-            days = "0"
-
-        time_parts = other_parts.split(":")
-        missing_parts = [0 for _ in range(3 - len(time_parts))]
-        seconds = time_parts.pop(-1)
-        seconds, _, microseconds = seconds.partition(".")
-        time_parts.extend([seconds, microseconds or "0"])
-        all_parts = map(int, (days, *missing_parts, *time_parts))
-        return Duration(*all_parts).as_timedelta().total_seconds()
-
-    def as_timedelta(self) -> timedelta:
-        return timedelta(**self._asdict())
-
-
-if TYPE_CHECKING:
-    from collections.abc import Generator, Mapping, Sequence
-
-    from cyberdrop_dl.managers.manager import Manager
-
-FFPROBE_CALL_PREFIX = "ffprobe", "-hide_banner", "-loglevel", "error", "-show_streams", "-print_format", "json"
-
-
-class StreamDict(TypedDict, total=False):
-    index: Required[int]
-    codec_type: Required[Literal["video", "audio", "subtitle"]]
-
-
-class FFprobeOutput(TypedDict):
-    streams: list[StreamDict]
-
-
-class Tags(CIMultiDict[Any]): ...
-
-
-class FPS(float):
-    def __str__(self) -> str:
-        return str(int(self)) if self.is_integer() else f"{self:.2f}"
-
-
-@dataclass(frozen=True, kw_only=True)
-class Stream:
-    index: int
-    codec_name: str
-    codec_type: str
-    bitrate: int | None
-    duration: float | None
-    tags: Tags
-
-    @property
-    def length(self):
-        return self.duration
-
-    @property
-    def codec(self) -> str:
-        return self.codec_name
-
-    @classmethod
-    def validate(cls, stream_info: Mapping[str, Any]) -> dict[str, Any]:
-        info = get_valid_dict(cls, stream_info)
-        tags = Tags(stream_info.get("tags", {}))
-        duration: float | str | None = stream_info.get("duration") or tags.get("duration")
-        bitrate = int(stream_info.get("bitrate") or stream_info.get("bit_rate") or 0) or None
-        duration = Duration.parse(duration) if duration else None
-        new_attrs = {"tags": tags, "duration": duration, "bitrate": bitrate}
-        return info | new_attrs
-
-    @classmethod
-    def new(cls, stream_info: Mapping[str, Any]) -> Self:
-        return cls(**cls.validate(stream_info))
-
-    def as_dict(self) -> dict[str, Any]:
-        return asdict(self) | {"tags": dict(self.tags)}
-
-
-@dataclass(frozen=True, kw_only=True)
-class AudioStream(Stream):
-    sample_rate: int | None
-    codec_type: Literal["audio"] = "audio"
-
-    @classmethod
-    def validate(cls, stream_info: StreamDict) -> dict[str, Any]:
-        sample_rate = int(float(stream_info.get("sample_rate", 0))) or None
-        defaults = super().validate(stream_info)
-        return defaults | {"sample_rate": sample_rate}
-
-
-@dataclass(frozen=True, kw_only=True)
-class VideoStream(Stream):
-    width: int | None
-    height: int | None
-    fps: FPS | None
-    resolution: str | None
-    codec_type: Literal["video"] = "video"
-
-    @classmethod
-    def validate(cls, stream_info: StreamDict) -> dict[str, Any]:
-        width = int(float(stream_info.get("width", 0))) or None
-        height = int(float(stream_info.get("height", 0))) or None
-        resolution = fps = None
-        if width and height:
-            resolution: str | None = f"{width}x{height}"
-
-        if (avg_fps := stream_info.get("avg_frame_rate")) and str(avg_fps) not in {"0/0", "0", "0.0"}:
-            fps: FPS | None = FPS(Fraction(avg_fps))
-
-        defaults = super().validate(stream_info)
-        new_attrs = {"width": width, "height": height, "fps": fps, "resolution": resolution}
-        return defaults | new_attrs
-
-
-@dataclass
-class FFprobeResult:
-    ffprobe_output: FFprobeOutput
-    streams: tuple[Stream, ...] = field(init=False)
-
-    def __post_init__(self):
-        streams: list[Stream] = []
-        for stream in self.ffprobe_output.get("streams", []):
-            if stream["codec_type"] == "video":
-                streams.append(VideoStream.new(stream))
-            elif stream["codec_type"] == "audio":
-                streams.append(AudioStream.new(stream))
-        self.streams = tuple(streams)
-
-    def video_streams(self) -> Generator[VideoStream]:
-        for stream in self.streams:
-            if isinstance(stream, VideoStream):
-                yield stream
-
-    def audio_streams(self) -> Generator[AudioStream]:
-        for stream in self.streams:
-            if isinstance(stream, AudioStream):
-                yield stream
-
-    @property
-    def audio(self) -> AudioStream:
-        """First audio stream"""
-        return next(self.audio_streams())
-
-    @property
-    def video(self) -> VideoStream:
-        """First video stream"""
-        return next(self.video_streams())
-
-    def __bool__(self) -> bool:
-        return bool(self.ffprobe_output.get("streams"))
-
-
-class FFprobe:
-    def __init__(self, manager: Manager | None = None) -> None:
-        self.manager = manager
-        self.version = ""
-        self.is_available = bool(self.version)
-
-    async def probe(self, file_or_url: Path | URL, /, *, headers: Mapping[str, str] | None = None) -> FFprobeResult:
-        if isinstance(file_or_url, URL):
-            assert file_or_url.absolute
-            assert file_or_url.scheme
-
-        elif isinstance(file_or_url, Path):
-            assert file_or_url.is_absolute()
-            assert not headers
-
-        else:
-            raise ValueError("Can only probe a Path or a yarl.URL")
-
-        command = *FFPROBE_CALL_PREFIX, str(file_or_url)
-        if headers:
-            add_headers = []
-            for name, value in headers.items():
-                add_headers.extend(["-headers", f"{name}: {value}"])
-
-            command = *command, *add_headers
-        result = await _run_command(command)
-        output = cast("FFprobeOutput", json.loads(result.stdout) if result.success else {"streams": []})
-        return FFprobeResult(output)
-
-
-class SubProcessResult(NamedTuple):
-    return_code: int | None
-    stdout: str
-    stderr: str
-    success: bool
-    command: tuple
-
-    def as_dict(self) -> dict:
-        joined_command = " ".join(self.command)
-        return self._asdict() | {"command": joined_command}
-
-
-async def _run_command(command: Sequence[str]) -> SubProcessResult:
-    joined_command = " ".join(command)
-    log_debug(f"Running ffprobe command: {joined_command}")
-    process = await asyncio.create_subprocess_exec(*command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    stdout, stderr = await process.communicate()
-    return_code = process.returncode
-    stdout_str = stdout.decode("utf-8", errors="ignore")
-    stderr_str = stderr.decode("utf-8", errors="ignore")
-    results = SubProcessResult(return_code, stdout_str, stderr_str, return_code == 0, tuple(command))
-    log_debug(results.as_dict())
-    return results
